snapshots: separate implementations into packages
