+++ snapshot/naive/naive.go
@@ -1,4 +1,4 @@
-package snapshot
+package naive
 
 import (
 	"fmt"
