+++ snapshot/naive/naive_test.go
@@ -1,4 +1,4 @@
-package snapshot
+package naive
 
 import (
 	"io/ioutil"
