+++ snapshot/testutil/helpers.go
@@ -1,4 +1,4 @@
-package snapshot
+package testutil
 
 import (
 	"os/exec"
@@ -7,13 +7,13 @@ import (
 	"github.com/docker/containerd"
 )
 
-func unmountAll(t *testing.T, mounts []containerd.Mount) {
+func UnmountAll(t *testing.T, mounts []containerd.Mount) {
 	for _, mount := range mounts {
-		unmount(t, mount.Target)
+		Unmount(t, mount.Target)
 	}
 }
 
-func unmount(t *testing.T, mountPoint string) {
+func Unmount(t *testing.T, mountPoint string) {
 	t.Log("unmount", mountPoint)
 	umount := exec.Command("umount", mountPoint)
 	err := umount.Run()
